test: hoist the shared manifest fixture helper into __utils__
